+++ sentry-android-replay/src/main/java/io/sentry/android/replay/capture/BufferCaptureStrategy.kt
@@ -20,6 +20,7 @@ import io.sentry.transport.ICurrentDateProvider
 import io.sentry.util.FileUtils
 import java.io.File
 import java.security.SecureRandom
+import java.util.Date
 import java.util.concurrent.ScheduledExecutorService
 
 internal class BufferCaptureStrategy(
@@ -34,39 +35,9 @@ internal class BufferCaptureStrategy(
     // TODO: capture envelopes for buffered segments instead, but don't send them until buffer is triggered
     private val bufferedSegments = mutableListOf<ReplaySegment.Created>()
 
-    // TODO: rework this bs, it doesn't work with sending replay on restart
-    private val bufferedScreensLock = Any()
-    private val bufferedScreens = mutableListOf<Pair<String, Long>>()
-
     internal companion object {
         private const val TAG = "BufferCaptureStrategy"
         private const val ENVELOPE_PROCESSING_DELAY: Long = 100L
-    }
-
-    override fun start(
-        recorderConfig: ScreenshotRecorderConfig,
-        segmentId: Int,
-        replayId: SentryId
-    ) {
-        super.start(recorderConfig, segmentId, replayId)
-
-        hub?.configureScope {
-            val screen = it.screen?.substringAfterLast('.')
-            if (screen != null) {
-                synchronized(bufferedScreensLock) {
-                    bufferedScreens.add(screen to dateProvider.currentTimeMillis)
-                }
-            }
-        }
-    }
-
-    override fun onScreenChanged(screen: String?) {
-        synchronized(bufferedScreensLock) {
-            val lastKnownScreen = bufferedScreens.lastOrNull()?.first
-            if (screen != null && lastKnownScreen != screen) {
-                bufferedScreens.add(screen to dateProvider.currentTimeMillis)
-            }
-        }
     }
 
     override fun pause() {
@@ -90,7 +61,7 @@ internal class BufferCaptureStrategy(
 
     override fun captureReplay(
         isTerminating: Boolean,
-        onSegmentSent: () -> Unit
+        onSegmentSent: (Date) -> Unit
     ) {
         val sampled = random.sample(options.experimental.sessionReplay.errorSampleRate)
 
@@ -121,8 +92,7 @@ internal class BufferCaptureStrategy(
                 // we only want to increment segment_id in the case of success, but currentSegment
                 // might be irrelevant since we changed strategies, so in the callback we increment
                 // it on the new strategy already
-                // TODO: also pass new segmentTimestamp to the new strategy
-                onSegmentSent()
+                onSegmentSent(segment.replay.timestamp)
             }
         }
     }
@@ -136,7 +106,7 @@ internal class BufferCaptureStrategy(
 
             val now = dateProvider.currentTimeMillis
             val bufferLimit = now - options.experimental.sessionReplay.errorReplayDuration
-            cache?.rotate(bufferLimit)
+            screenAtStart = cache?.rotate(bufferLimit)
             bufferedSegments.rotate(bufferLimit)
         }
     }
@@ -159,7 +129,7 @@ internal class BufferCaptureStrategy(
         }
         // we hand over replayExecutor to the new strategy to preserve order of execution
         val captureStrategy = SessionCaptureStrategy(options, hub, dateProvider, replayExecutor)
-        captureStrategy.start(recorderConfig, segmentId = currentSegment, replayId = currentReplayId)
+        captureStrategy.start(recorderConfig, segmentId = currentSegment, replayId = currentReplayId, replayType = BUFFER)
         return captureStrategy
     }
 
@@ -167,21 +137,6 @@ internal class BufferCaptureStrategy(
         super.onTouchEvent(event)
         val bufferLimit = dateProvider.currentTimeMillis - options.experimental.sessionReplay.errorReplayDuration
         rotateEvents(currentEvents, bufferLimit)
-    }
-
-    private fun findAndSetStartScreen(segmentStart: Long) {
-        synchronized(bufferedScreensLock) {
-            val startScreen = bufferedScreens.lastOrNull { (_, timestamp) ->
-                timestamp <= segmentStart
-            }?.first
-            // if no screen is found before the segment start, this likely means the buffer is from the
-            // app start, and the start screen will be taken from the navigation crumbs
-            if (startScreen != null) {
-                screenAtStart = startScreen
-            }
-            // can clear as we switch to session mode and don't care anymore about buffering
-            bufferedScreens.clear()
-        }
     }
 
     private fun deleteFile(file: File?) {
@@ -246,11 +201,9 @@ internal class BufferCaptureStrategy(
         val height = this.recorderConfig.recordingHeight
         val width = this.recorderConfig.recordingWidth
 
-        findAndSetStartScreen(currentSegmentTimestamp.time)
-
         replayExecutor.submitSafely(options, "$TAG.$taskName") {
             val segment =
-                createSegmentInternal(duration, currentSegmentTimestamp, replayId, segmentId, height, width, BUFFER)
+                createSegmentInternal(duration, currentSegmentTimestamp, replayId, segmentId, height, width)
             onSegmentCreated(segment)
         }
     }